+++ pyproj/database.pyx
@@ -7,7 +7,7 @@ from typing import Optional
 from libc.stdlib cimport free, malloc
 
 from pyproj._compat cimport cstrdecode, cstrencode
-from pyproj._datadir cimport pyproj_context_create, pyproj_context_destroy
+from pyproj._context cimport pyproj_context_create
 
 from pyproj.aoi import AreaOfUse
 from pyproj.enums import PJType
@@ -71,7 +71,6 @@ def get_authorities():
     cdef PJ_CONTEXT* context = pyproj_context_create()
     cdef PROJ_STRING_LIST proj_auth_list = proj_get_authorities_from_database(context)
     if proj_auth_list == NULL:
-        pyproj_context_destroy(context)
         return []
     cdef int iii = 0
     try:
@@ -80,7 +79,6 @@ def get_authorities():
             auth_list.append(proj_auth_list[iii])
             iii += 1
     finally:
-        pyproj_context_destroy(context)
         proj_string_list_destroy(proj_auth_list)
     return auth_list
 
@@ -105,19 +103,14 @@ def get_codes(str auth_name not None, pj_type not None, bint allow_deprecated=Fa
     list[str]:
         Codes associated with authorities in PROJ database.
     """
-    cdef PJ_CONTEXT* context = NULL
     cdef PJ_TYPE cpj_type = get_pj_type(pj_type)
     cdef PROJ_STRING_LIST proj_code_list = NULL
-    try:
-        context = pyproj_context_create()
-        proj_code_list = proj_get_codes_from_database(
-            context,
-            cstrencode(auth_name),
-            cpj_type,
-            allow_deprecated,
-        )
-    finally:
-        pyproj_context_destroy(context)
+    proj_code_list = proj_get_codes_from_database(
+        pyproj_context_create(),
+        cstrencode(auth_name),
+        cpj_type,
+        allow_deprecated,
+    )
     if proj_code_list == NULL:
         return []
     cdef int iii = 0
@@ -203,7 +196,6 @@ def query_crs_info(
     list[CRSInfo]:
         CRS information from the PROJ database.
     """
-    cdef PJ_CONTEXT* context = NULL
     cdef PJ_TYPE *pj_type_list = NULL
     cdef PROJ_CRS_LIST_PARAMETERS *query_params = NULL
     cdef PROJ_CRS_INFO **crs_info_list = NULL
@@ -226,7 +218,6 @@ def query_crs_info(
             for iii in range(pj_type_count):
                 pj_type_list[iii] = get_pj_type(pj_types[iii])
 
-        context = pyproj_context_create()
         query_params = proj_get_crs_list_parameters_create()
         query_params.types = pj_type_list
         query_params.typesCount = pj_type_count
@@ -240,7 +231,7 @@ def query_crs_info(
             query_params.north_lat_degree = area_of_interest.north_lat_degree
 
         crs_info_list = proj_get_crs_info_list_from_database(
-            context,
+            pyproj_context_create(),
             c_auth_name,
             query_params,
             &result_count)
@@ -249,7 +240,6 @@ def query_crs_info(
             proj_get_crs_list_parameters_destroy(query_params)
         if pj_type_list != NULL:
             free(pj_type_list)
-        pyproj_context_destroy(context)
     if crs_info_list == NULL:
         return []
     try:
@@ -425,7 +415,6 @@ def get_units_map(str auth_name=None, str category=None, bint allow_deprecated=F
             )
     finally:
         proj_unit_list_destroy(db_unit_list)
-        pyproj_context_destroy(context)
     return units_map
 
 
@@ -462,15 +451,11 @@ def get_database_metadata(str key not None):
     Optional[str]:
         The metatada information if available.
     """
-    cdef PJ_CONTEXT* context = pyproj_context_create()
     cdef const char* metadata = NULL
-    try:
-        metadata = proj_context_get_database_metadata(
-            context,
-            cstrdecode(key),
-        )
-        if metadata == NULL:
-            return None
-        return metadata
-    finally:
-        pyproj_context_destroy(context)
+    metadata = proj_context_get_database_metadata(
+        pyproj_context_create(),
+        cstrdecode(key),
+    )
+    if metadata == NULL:
+        return None
+    return metadata