+++ test/test_network.py
@@ -7,7 +7,7 @@ from pyproj.network import set_ca_bundle_path
 
 
 @patch.dict("os.environ", {}, clear=True)
-@patch("pyproj.network._set_ca_bundle_path")
+@patch("pyproj.network._set_context_ca_bundle_path")
 def test_ca_bundle_path__default(c_set_ca_bundle_path_mock):
     set_ca_bundle_path()
     c_set_ca_bundle_path_mock.assert_called_with(certifi.where())
@@ -16,7 +16,7 @@ def test_ca_bundle_path__default(c_set_ca_bundle_path_mock):
 @pytest.mark.parametrize(
     "env_var", ["PROJ_CURL_CA_BUNDLE", "CURL_CA_BUNDLE", "SSL_CERT_FILE"]
 )
-@patch("pyproj.network._set_ca_bundle_path")
+@patch("pyproj.network._set_context_ca_bundle_path")
 def test_ca_bundle_path__always_certifi(c_set_ca_bundle_path_mock, env_var):
     with patch.dict("os.environ", {env_var: "/tmp/dummy/path/cacert.pem"}, clear=True):
         set_ca_bundle_path(True)
@@ -24,7 +24,7 @@ def test_ca_bundle_path__always_certifi(c_set_ca_bundle_path_mock, env_var):
 
 
 @patch.dict("os.environ", {}, clear=True)
-@patch("pyproj.network._set_ca_bundle_path")
+@patch("pyproj.network._set_context_ca_bundle_path")
 def test_ca_bundle_path__skip(c_set_ca_bundle_path_mock):
     set_ca_bundle_path(False)
     c_set_ca_bundle_path_mock.assert_called_with("")
@@ -33,7 +33,7 @@ def test_ca_bundle_path__skip(c_set_ca_bundle_path_mock):
 @pytest.mark.parametrize(
     "env_var", ["PROJ_CURL_CA_BUNDLE", "CURL_CA_BUNDLE", "SSL_CERT_FILE"]
 )
-@patch("pyproj.network._set_ca_bundle_path")
+@patch("pyproj.network._set_context_ca_bundle_path")
 def test_ca_bundle_path__env_var_skip(c_set_ca_bundle_path_mock, env_var):
     with patch.dict("os.environ", {env_var: "/tmp/dummy/path/cacert.pem"}, clear=True):
         set_ca_bundle_path()
@@ -43,7 +43,7 @@ def test_ca_bundle_path__env_var_skip(c_set_ca_bundle_path_mock, env_var):
 @pytest.mark.parametrize(
     "env_var", ["PROJ_CURL_CA_BUNDLE", "CURL_CA_BUNDLE", "SSL_CERT_FILE"]
 )
-@patch("pyproj.network._set_ca_bundle_path")
+@patch("pyproj.network._set_context_ca_bundle_path")
 def test_ca_bundle_path__custom_path(c_set_ca_bundle_path_mock, env_var):
     with patch.dict("os.environ", {env_var: "/tmp/dummy/path/cacert.pem"}, clear=True):
         set_ca_bundle_path("/my/path/to/cacert.pem")