+++ test/test_transformer.py
@@ -1140,9 +1140,6 @@ def test_transformer_group__download_grids__directory(
             )
 
 
-@pytest.mark.skipif(
-    pyproj._datadir._USE_GLOBAL_CONTEXT, reason="Global Context not Threadsafe."
-)
 def test_transformer_multithread__pipeline():
     # https://github.com/pyproj4/pyproj/issues/782
     trans = Transformer.from_pipeline(
@@ -1158,9 +1155,6 @@ def test_transformer_multithread__pipeline():
             pass
 
 
-@pytest.mark.skipif(
-    pyproj._datadir._USE_GLOBAL_CONTEXT, reason="Global Context not Threadsafe."
-)
 def test_transformer_multithread__crs():
     # https://github.com/pyproj4/pyproj/issues/782
     trans = Transformer.from_crs(4326, 3857)