+++ docs/history.rst
@@ -6,6 +6,7 @@ Latest
 - DEP: Minimum supported Python version 3.9 (issue #1111)
 - ENH: Added allow_superseded kwargs to :class:`pyproj.transformer.TransformerGroup` (pull #1269)
 - ENH: Added :meth:`CRS.to_2d` to demote 3D CRS to 2D (issue #1266)
+- PERF: thread local context (issue #1133)
 
 3.5.0
 ------